pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neurosym-amex"
version = "0.1.0"
description = "Neuro-Symbolic Decision Engine (Amex-like Flows)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "z3-solver>=4.12.2.0",
  "pydantic>=2.6",
  "pyyaml>=6.0.1",
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "pytest>=8.0"
]

[tool.setuptools.packages.find]
where = ["src"]

